fix: coding style

+++ FFT.cpp
@@ -2,14 +2,14 @@
 
 vector<int16_t> paddingToPowerOfTwo(vector<int16_t> data)
 {
-    unsigned int n = data.size();
+    size_t n = data.size();
 
     // if power is 2 then return
     if ((n & (n - 1)) == 0)
         return data;
 
     // get the next number which power is 2
-    unsigned int base_2_num = 1;
+    size_t base_2_num = 1;
     while (base_2_num < n)
         base_2_num <<= 1;
